+++ pygubu/builder/tkstdwidgets.py
@@ -1,12 +1,6 @@
 # encoding: utf8
-from __future__ import unicode_literals
-
 import logging
-
-try:
-    import tkinter as tk
-except ImportError:
-    import Tkinter as tk
+import tkinter as tk
 
 from .builderobject import (BuilderObject, EntryBaseBO, PanedWindowBO,
                             PanedWindowPaneBO, register_widget, CB_TYPES)
@@ -22,6 +16,7 @@ class TKToplevel(BuilderObject):
     class_ = tk.Toplevel
     container = True
     layout_required = False
+    container_layout = True
     allowed_parents = ('root',)
     # maxchildren = 2  # A menu and a frame
     OPTIONS_STANDARD = ('borderwidth', 'cursor', 'highlightbackground',
@@ -49,9 +44,9 @@ class TKToplevel(BuilderObject):
             self.widget = self.class_(master, **args)
         return self.widget
 
-    def layout(self, target=None):
+    # def layout(self, target=None):
         # we marked this widget as not allowed to edit layoutu
-        pass
+    #    pass
 
     def _set_property(self, target_widget, pname, value):
         method_props = ('geometry', 'overrideredirect', 'title')
@@ -118,6 +113,7 @@ class TKFrame(BuilderObject):
                         'height', 'width')
     class_ = tk.Frame
     container = True
+    container_layout = True
     properties = OPTIONS_STANDARD + OPTIONS_SPECIFIC
     ro_properties = ('class_', 'container')
 
@@ -146,6 +142,7 @@ register_widget('tk.Label', TKLabel, 'Label', ('Control & Display', 'tk'))
 class TKLabelFrame(BuilderObject):
     class_ = tk.LabelFrame
     container = True
+    container_layout = True
     OPTIONS_STANDARD = (
         'borderwidth', 'cursor', 'font', 'foreground',
         'highlightbackground', 'highlightcolor', 'highlightthickness',
@@ -911,7 +908,7 @@ class TKLabelwidgetBO(BuilderObject):
     def add_child(self, bobject):
         self.widget.configure(labelwidget=bobject.widget)
 
-    def layout(self, target=None, configure_gridrc=True):
+    def layout(self, target=None):
         pass
 
     #
@@ -952,7 +949,7 @@ class ToplevelMenuHelperBO(BuilderObject):
     def add_child(self, bobject):
         self.widget.configure(menu=bobject.widget)
 
-    def layout(self, target=None, configure_gridrc=True):
+    def layout(self, target=None):
         pass
 
     #