+++ pygubu/widgets/dialog.py
@@ -1,12 +1,6 @@
 # encoding: UTF-8
-from __future__ import unicode_literals
-
-try:
-    import tkinter as tk
-    import tkinter.ttk as ttk
-except ImportError:
-    import Tkinter as tk
-    import ttk
+import tkinter as tk
+import tkinter.ttk as ttk
 
 
 class Dialog(object):